+++ backend/ARCHITECTURE.rst
@@ -3,9 +3,9 @@
 
     *   - Codebase
         - Description
-    *   - `Barista`_
+    *   - `Tsukaeru`_
         - AioHTTP ( Proxy | Media ) 
     
 
 
-.. _Barista: https://github.com/baseplate-admin/CoreProject/tree/master/backend/butler
+.. _Tsukaeru: https://github.com/baseplate-admin/CoreProject/tree/master/backend/tsukaeru
