+++ backend/run_aiohttp.py
@@ -1,4 +1,4 @@
-from barista.server import aiohttp_app
+from tsukaeru.server import aiohttp_app
 
 from aiohttp import web
